feat: basic migration of appView
