+++ migration-to-ts/src/components/view/sources/sources.ts
@@ -1,22 +1,22 @@
 import './sources.css';
-import { ISourse } from '../../../types/index';
+import { ISources } from '../../../types/index';
 
 class Sources {
-    draw(data: ISourse[]) {
-        const fragment = document.createDocumentFragment() as DocumentFragment;
-        const sourceItemTemp = document.querySelector('#sourceItemTemp') as HTMLTemplateElement;
+  draw(data: ISources[]) {
+    const fragment = document.createDocumentFragment() as DocumentFragment;
+    const sourceItemTemp = document.querySelector('#sourceItemTemp') as HTMLTemplateElement;
 
-        data.forEach((item) => {
-            const sourceClone = sourceItemTemp.content.cloneNode(true) as HTMLTemplateElement;
+    data.forEach((item) => {
+      const sourceClone = sourceItemTemp.content.cloneNode(true) as HTMLTemplateElement;
 
-            (sourceClone.querySelector('.source__item-name') as HTMLDivElement).textContent = item.name;
-            (sourceClone.querySelector('.source__item') as HTMLDivElement).setAttribute('data-source-id', item.id);
+      (sourceClone.querySelector('.source__item-name') as HTMLDivElement).textContent = item.name;
+      (sourceClone.querySelector('.source__item') as HTMLDivElement).setAttribute('data-source-id', item.id);
 
-            fragment.append(sourceClone);
-        });
+      fragment.append(sourceClone);
+    });
 
-        (document.querySelector('.sources') as HTMLDivElement).append(fragment);
-    }
+    (document.querySelector('.sources') as HTMLDivElement).append(fragment);
+  }
 }
 
 export default Sources;
