--- migration-to-ts/types.ts
@@ -1,1 +1,0 @@
-const a: string = 'gg';
